# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: M&S Delivery System API
  version: 1.0.0
  description: Official REST API specification for M&S Delivery System (FastAPI backend on Vercel).
servers:
  - url: /api/v1
    description: Current API v1 base path
paths:
  /auth/register:
    post:
      summary: Register a new customer
      operationId: registerCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRegisterRequest'
      responses:
        '201':
          description: Customer successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /auth/login:
    post:
      summary: Login user (Customer, Rider, or Admin)
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /auth/refresh:
    post:
      summary: Refresh access token using rotating refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /auth/logout:
    post:
      summary: Revoke session and logout
      operationId: logoutUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /auth/me:
    get:
      summary: Get current authenticated user details
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
  /customers/addresses:
    get:
      summary: List saved addresses for current customer
      operationId: listCustomerAddresses
      security:
        - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
    post:
      summary: Save a new delivery address
      operationId: createCustomerAddress
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressCreateRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
...
  /menu-items:
    post:
      summary: Create a menu item (Admin only)
      operationId: createMenuItem
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuItemCreateRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItem'
  /menu-items/{item_id}:
    put:
      summary: Update menu item (Admin only)
      operationId: updateMenuItem
      security:
        - BearerAuth: []
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuItemUpdateRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItem'
    delete:
      summary: Delete menu item (Admin only)
      operationId: deleteMenuItem
      security:
        - BearerAuth: []
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ErrorResponse:
      description: Standard error envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
  schemas:
...
    UserRole:
      type: string
      enum: [customer, rider, admin]
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    CustomerRegisterRequest:
      type: object
      required: [email, password, full_name, phone]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        full_name:
          type: string
        phone:
          type: string
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    AuthTokenResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in, user]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          example: 3600
        user:
          $ref: '#/components/schemas/UserSummary'
    UserSummary:
      type: object
      required: [id, email, full_name, role]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        full_name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        phone:
          type: string
    UserDetailResponse:
      type: object
      required: [id, email, full_name, role, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        full_name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        phone:
          type: string
        created_at:
          type: string
          format: date-time
    Address:
      type: object
      required: [id, label, address_line, latitude, longitude, is_default]
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
          example: Home
        address_line:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        delivery_notes:
          type: string
        is_default:
          type: boolean
    AddressCreateRequest:
      type: object
      required: [label, address_line, latitude, longitude]
      properties:
        label:
          type: string
        address_line:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        delivery_notes:
          type: string
        is_default:
          type: boolean
          default: false
...
    MenuItem:
      type: object
      required: [id, store_id, name, price, is_available]
      properties:
        id:
          type: string
          format: uuid
        store_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        category_name:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        image_url:
          type: string
        is_available:
          type: boolean
    MenuItemCreateRequest:
      type: object
      required: [store_id, name, price]
      properties:
        store_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        price:
          type: number
        image_url:
          type: string
        is_available:
          type: boolean
          default: true
    MenuItemUpdateRequest:
      type: object
      properties:
        category_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        price:
          type: number
        image_url:
          type: string
        is_available:
          type: boolean
...